requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
requests
openai
pydantic-settings
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
icalendar